Feature/table

+++ backend/controllers/auth.controllers.js
@@ -80,6 +80,16 @@ const login = async (req, res, next) => {
     }   
 };
 
+const logout = async(req, res, next) => {
+    res.clearCookie('token', {
+        httpOnly: true,
+        secure: process.env.NODE_ENV === 'production',
+        sameSite: process.env.NODE_ENV === 'production' ? 'none' : 'lax',
+        maxAge: 24 * 60 * 60 * 1000
+    });
+    return res.status(200).json({message: "Log out successful"});
+} 
+
 const verifyEmail = async (req, res, next) => {
     const { token } = req.query;
 
@@ -105,13 +115,19 @@ const verifyEmail = async (req, res, next) => {
             
             return res.status(400).json({message: "The token was expired. New token has been send to your email"});
         }
+        
+        const user_status = await db.one('SELECT status FROM users WHERE id = $1', [dbToken.user_id]); 
+        if (user_status.status === 'blocked') {
+            return res.status(403).send('<p>You have been blocked.</p>')
+        }
+
 
         await db.tx(async t => {
             await t.none('DELETE FROM email_verifications WHERE token = $1', [token]);
             await t.none("UPDATE users SET status = 'verified' WHERE id = $1", [dbToken.user_id]);
         });
         
-        return res.status(200).send("<h1>Email has been verified. You can return to the application</h1>"});
+        return res.status(200).send("<p>Email has been verified. You can return to the application</p>");
     } catch (error) {
         next(error);
     }
@@ -133,5 +149,6 @@ async function sendEmailVerificationToken(txClient, userId, targetEmail) {
 module.exports = {
     register,
     login,
-    verifyEmail
+    verifyEmail,
+    logout,
 };
